Fix Particles being emissive, added a check for particle spawning
