+++ src/main/java/me/hydos/lint/client/LintClient.java
@@ -43,7 +43,7 @@ import me.hydos.lint.entity.Entities;
 import me.hydos.lint.fluid.LintFluids;
 import me.hydos.lint.network.ClientNetworking;
 import me.hydos.lint.network.Networking;
-import me.hydos.lint.particle.FallenMysticalLeafAnimated;
+import me.hydos.lint.particle.FallenMysticalLeaf;
 import me.hydos.lint.particle.Particles;
 import me.hydos.lint.screenhandler.ScreenHandlers;
 import me.hydos.lint.screenhandler.client.LilTaterScreen;
@@ -95,7 +95,7 @@ public class LintClient implements ClientModInitializer {
 	}
 
 	private void registerParticles() {
-		ParticleFactoryRegistry.getInstance().register(Particles.FALLEN_MYSTICAL_LEAF, FallenMysticalLeafAnimated.FallenMysticalLeaf.Factory::new);
+		ParticleFactoryRegistry.getInstance().register(Particles.FALLEN_MYSTICAL_LEAF, FallenMysticalLeaf.Factory::new);
 	}
 
 	private void registerBossMusicFixes() {
