+++ src/main/java/me/hydos/lint/particle/FallenMysticalLeaf.java
@@ -1,0 +1,69 @@
+/*
+ * Lint
+ * Copyright (C) 2020 hYdos, Valoeghese, ramidzkh
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
+
+package me.hydos.lint.particle;
+
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
+import net.fabricmc.fabric.api.client.particle.v1.FabricSpriteProvider;
+import net.minecraft.client.particle.*;
+import net.minecraft.client.render.WorldRenderer;
+import net.minecraft.client.world.ClientWorld;
+import net.minecraft.particle.DefaultParticleType;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.World;
+import org.jetbrains.annotations.Nullable;
+
+@Environment(EnvType.CLIENT)
+public class FallenMysticalLeaf extends AnimatedParticle {
+
+    protected FallenMysticalLeaf(World world, double x, double y, double z, SpriteProvider sprites) {
+        super((ClientWorld) world, x, y, z, sprites, -3.0E-3F);
+        this.age = 0;
+        this.maxAge = 80;
+        setSprite(sprites.getSprite(world.random));
+    }
+
+    @Override
+    public ParticleTextureSheet getType() {
+        return ParticleTextureSheet.PARTICLE_SHEET_OPAQUE;
+    }
+
+    @Override
+    public int getColorMultiplier(float tint) {
+        BlockPos blockPos = new BlockPos(this.x, this.y, this.z);
+        return this.world.isChunkLoaded(blockPos) ? WorldRenderer.getLightmapCoordinates(this.world, blockPos) : 0;
+    }
+
+    @Environment(EnvType.CLIENT)
+    public static class Factory implements ParticleFactory<DefaultParticleType> {
+
+        private final FabricSpriteProvider sprites;
+
+        public Factory(FabricSpriteProvider sprites) {
+            this.sprites = sprites;
+        }
+
+        @Nullable
+        @Override
+        public Particle createParticle(DefaultParticleType parameters, ClientWorld world, double x, double y, double z, double velocityX, double velocityY, double velocityZ) {
+            return new FallenMysticalLeaf(world, x, y, z, this.sprites);
+        }
+    }
+}
